Move DOM Manipulation to the domupdates.js file
