+++ src/scripts.js
@@ -11,37 +11,20 @@ import './images/turing-logo.png'
 console.log('This is the JavaScript entry file - your code begins here.');
 
 
-var dateControl = document.querySelector('input[type="date"]');
-dateControl.value
-console.log(dateControl.value);
-console.log(dateControl.valueAsNumber)
 
-import Customer from './classes/customer';
-import customerData from '../test/test-data/customer-data'
-import bookingData from '../test/test-data/booking-data'
-import roomData from '../test/test-data/room-data'
-const welcomeMessage = document.querySelector('.welcome-user');
-const bookingInfoPage = document.querySelector('.dashboard')
 
-window.addEventListener('load', welcomeUserMessage())
+import domUpdates from './domUpdates';
 
-function welcomeUserMessage() {
-  const customer = new Customer(customerData[3]);
-  customer.currentBookings(bookingData)
-  const totalCost = customer.calculateTotalCost(roomData)
-  console.log(totalCost)
-  welcomeMessage.innerHTML = `<h2>Welcome To The Overlook Hotel ${customer.name}</h2> 
-  <h2>You Spent: $ ${totalCost} So Far On Rooms!</h2>`
+import { calendarSubmitBtn } from './domUpdates';
 
-console.log(customer.presentBookings)
 
-  customer.presentBookings.forEach(booking => {
-    bookingInfoPage.innerHTML += `<section>
-    <h2>Booking Information:</h2>
-    <h2>booking id: ${booking.id}</h2>
-    <h2>Booked for: ${booking.date}<h2>
-    <h2>Room Number: ${booking.roomNumber}</h2>
-    </section>`
-  })
-}
+
+window.addEventListener('load', domUpdates.welcomeUserMessage())
+
+calendarSubmitBtn.addEventListener('click', function(event) {
+  domUpdates.grabdate(event)
+})
+
+
+
